Move create topic logic into root class where container was
